+++ Tests/Extends/v5.14.14.prefixes/1.build.ps1
@@ -1,0 +1,9 @@
+
+param(
+	[ValidateScript({"B2::B2.build.ps1"})]
+	$Extends
+)
+
+task main meta, B2::build2
+
+task . main