+++ Tests/Extends/v5.14.14.prefixes/B1.build.ps1
@@ -1,0 +1,11 @@
+
+task meta {
+}
+
+task build meta, {
+	'Building...'
+}
+
+task test {
+	'Do some tests.'
+}