+++ Tests/Extends/v5.14.14.prefixes/B2.build.ps1
@@ -1,0 +1,11 @@
+
+param(
+	[ValidateScript({"B1.build.ps1"})]
+	$Extends
+)
+
+task build2 build, test2
+
+task test2 test, {
+	'Do more tests.'
+}